example/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-project-api"
version = "0.1.0"
description = "Python web application built with lessweb framework for task management"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.9.1",
    "aiojobs",
    "toml",
    "orjson",
    "typing_inspect",
    "pydantic",
    "python-dotenv",
    "cryptography",
    "mysql-connector",
    "aiomysql",
    "commondao>=1.3.0",
    "pyway",
    "bullmq",
    "redis",
    "lessweb>=2.1.1",
    "PyJWT",
    "aiohttp-middlewares",
]

[project.optional-dependencies]
test = [
    "flake8",
    "mypy",
    "coverage",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-asyncio",
    "pytest-aiohttp",
    "types-toml",
    "types-retry",
    "types-PyMySQL",
]

[tool.mypy]
implicit_optional = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]